+++ app/models/peep.rb
@@ -1,0 +1,8 @@
+class Peep
+  include DataMapper::Resource
+
+  belongs_to :user
+
+  property :id, Serial
+  property :content, Text
+end
