+++ app/models/user.rb
@@ -12,6 +12,8 @@ class User
   property :username, String
   property :password_digest, Text
 
+  has n, :peeps
+
   def password=(password)
     @password = password
     self.password_digest = BCrypt::Password.create(password)
