+++ app/views/index.erb
@@ -1,1 +1,3 @@
 Welcome to Chitter
+
+<a href="/peeps/new">New peep</a>
