+++ app/views/peeps/index.erb
@@ -1,0 +1,4 @@
+<% @peeps.each do |peep| %>
+  <%= peep.content %>
+  <%= peep.user.name %>
+<% end %>
