+++ app/views/peeps/new.erb
@@ -1,0 +1,4 @@
+<form method="post" action="/peeps">
+  <input type="text" name="peep" value="" />
+  <input type="submit" value="Submit" />
+</form>
