+++ spec/features/user_management_spec.rb
@@ -40,22 +40,4 @@ feature "User management" do
     visit "/peeps/new"
     expect(page).to have_content("You must be signed in to post peeps")
   end
-
-  def create_user(email: "stefan@makersacademy.com", password: "supersecret")
-    User.create(email: email,
-                password: password,
-                password_confirmation: password,
-                name: "Stefan",
-                username: "sdorunga")
-  end
-
-  def sign_in
-    user = create_user(password: "password")
-    visit "/"
-    click_link "Sign In"
-    fill_in :email, with: user.email
-    fill_in :password, with: "password"
-    click_button "Sign In"
-  end
-
 end
